Creata activity astratta separata per i dialog degli errori di rete
